backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "mcp>=1.27.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.26",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]
